events move to regular sequelize models
