+++ core/api/__tests__/actions/events.ts
@@ -223,87 +223,5 @@ describe("actions/events", () => {
       expect(error).toBeUndefined();
       expect(success).toBe(true);
     });
-
-    describe("destroy multiple events", () => {
-      beforeEach(async () => {
-        await api.events.model.destroyFor();
-
-        await helper.factories.event({
-          profileGuid: "abc",
-          type: "pageview",
-          occurredAt: new Date(0),
-        });
-        await helper.factories.event({ profileGuid: "abc", type: "pageview" });
-        await helper.factories.event({ profileGuid: "abc", type: "purchase" });
-        await helper.factories.event({
-          profileGuid: "def",
-          type: "pageview",
-          occurredAt: new Date(0),
-        });
-        await helper.factories.event({ profileGuid: "def", type: "pageview" });
-        await helper.factories.event({ profileGuid: "def", type: "purchase" });
-      });
-
-      test("an administrator can destroy events for a profile", async () => {
-        connection.params = {
-          csrfToken,
-          profileGuid: "abc",
-        };
-        const { error, count } = await specHelper.runAction(
-          "events:destroy",
-          connection
-        );
-        expect(error).toBeFalsy();
-        expect(count).toBe(3);
-        const remainingEvents = await api.events.model.count();
-        expect(remainingEvents).toBe(3);
-      });
-
-      test("an administrator can destroy events for a topic", async () => {
-        connection.params = {
-          csrfToken,
-          type: "pageview",
-        };
-        const { error, count } = await specHelper.runAction(
-          "events:destroy",
-          connection
-        );
-        expect(error).toBeFalsy();
-        expect(count).toBe(4);
-        const remainingEvents = await api.events.model.count();
-        expect(remainingEvents).toBe(2);
-      });
-
-      test("an administrator can destroy events for a profile + topic", async () => {
-        connection.params = {
-          csrfToken,
-          profileGuid: "abc",
-          type: "pageview",
-        };
-        const { error, count } = await specHelper.runAction(
-          "events:destroy",
-          connection
-        );
-        expect(error).toBeFalsy();
-        expect(count).toBe(2);
-        const remainingEvents = await api.events.model.count();
-        expect(remainingEvents).toBe(4);
-      });
-
-      test("an administrator can destroy events before a date", async () => {
-        connection.params = {
-          csrfToken,
-          before: new Date(1000).getTime(),
-        };
-        const { error, count } = await specHelper.runAction(
-          "events:destroy",
-          connection
-        );
-        expect(error).toBeFalsy();
-        expect(count).toBe(2);
-        const remainingEvents = await api.events.model.count();
-        expect(remainingEvents).toBe(4);
-      });
-    });
   });
 });
