+++ core/api/src/index.ts
@@ -8,6 +8,8 @@ export { ApiKey } from "./models/ApiKey";
 export { Destination } from "./models/Destination";
 export { DestinationGroup } from "./models/DestinationGroup";
 export { DestinationGroupMembership } from "./models/DestinationGroupMembership";
+export { Event } from "./models/Event";
+export { EventData } from "./models/EventData";
 export { Export } from "./models/Export";
 export { File } from "./models/File";
 export { Group } from "./models/Group";
@@ -28,9 +30,3 @@ export { Setting } from "./models/Setting";
 export { Source } from "./models/Source";
 export { Team } from "./models/Team";
 export { TeamMember } from "./models/TeamMember";
-export {
-  EventArgs,
-  EventDataArgs,
-  EventPrototype,
-  EventDataPrototype,
-} from "./classes/events";
